fix dft logic

+++ Core/Src/magnetic_field.c
@@ -1,44 +1,45 @@
 #include "magnetic_field.h"
 
 MagneticField sampleMagneticField(
-		HAL_StatusTypeDef (*readMagneticSensor)(I2C_HandleTypeDef*, double*),
+		HAL_StatusTypeDef (*readMagneticSensor)(I2C_HandleTypeDef*, float*),
 		I2C_HandleTypeDef *i2c) {
 	MagneticField mf_sample;
 
-	double d_response[3];
+	float d_response[3];
 
 	if (HAL_OK == readMagneticSensor(i2c, d_response)) {
 
 		mf_sample.x = d_response[0];
 		mf_sample.y = d_response[1];
 		mf_sample.z = d_response[2];
+
 	}
 
 	return mf_sample;
 }
 
-void identifyMagneticField(double complex *d_mf_x_dft, double complex *d_mf_y_dft, double complex *d_mf_z_dft,
-		MagneticFieldSource *mf_nodes) {
+void identifyMagneticField(float complex *d_mf_x_dft, float complex *d_mf_y_dft,
+		float complex *d_mf_z_dft, MagneticFieldSource *mf_nodes) {
 
-	for(int i = 0; i < 3; i++){
+	for (int i = 0; i < 3; i++) {
 		int bin = (int) mf_nodes[i].i_frequency / 2;
 
-		double complex value = d_mf_z_dft[bin];
+		float complex value = d_mf_z_dft[bin];
 
 		mf_nodes[i].mf_intensity.z = cabs(value);
 	}
 }
 
-double getDistanceFromRSS(MagneticFieldSource node) {
-	double distance = cbrt(node.d_magnetic_cte / node.mf_intensity.z);
+float getDistanceFromRSS(MagneticFieldSource node) {
+	float distance = cbrt(node.d_magnetic_cte / node.mf_intensity.z);
 
 	return distance;
 }
 
-SpacePosition estimatePoisition(MagneticFieldSource *nodes ) {
+SpacePosition estimatePoisition(MagneticFieldSource *nodes) {
 	SpacePosition sp_rx_position;
 
-	double d_distances[3];
+	float d_distances[3];
 
 	for (int nodeIndex = 0; nodeIndex < 3; nodeIndex++) {
 		d_distances[nodeIndex] = getDistanceFromRSS(nodes[nodeIndex]);
